+++ src/Application/View/Participant/CardView.php
@@ -57,6 +57,8 @@ class CardView
     /** @var bool */
     public $isCheckedToday;
 
+    public ?bool $isOnline;
+
     /**
      * @param int    $id
      * @param bool   $editable
@@ -79,7 +81,8 @@ class CardView
         $owner,
         $sheetId,
         bool $getCheckinStatus = false,
-        bool $isCheckedToday = false
+        bool $isCheckedToday = false,
+        ?bool $isOnline = null
     ) {
         $this->id = $id;
         $this->editable = $editable;
@@ -92,5 +95,6 @@ class CardView
         $this->initials = (new GetParticipantInitials())($firstname, $lastname);
         $this->getCheckinStatus = $getCheckinStatus;
         $this->isCheckedToday = $isCheckedToday;
+        $this->isOnline = $isOnline;
     }
 }