+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2393 - Meet Online button
+
 ## [2.94.3] - 2021-06-08
 ### Fixed
 - MV-339 - Display call visio button in the networking private chat